requirements-dev.txt
pytest
pytest-mock
python-dotenv
requests
black
flake8
pandas
openpyxl
openai
python-dotenv